Calculate the additive function value for edges in a LSN

Calculate the additive function value for each edge feature in a Landscape Network (LSN)

afv_edges(
  edges,
  lsn_path,
  infl_col,
  segpi_col,
  afv_col,
  save_local = TRUE,
  overwrite = TRUE
)

Arguments

edges

An sf object with LINESTING geometry created using lines_to_lsn.

lsn_path

Local pathname to a directory in character format specifying where relationships.csv resides, which is created using link[SSNbler]{lines_to_lsn}.

infl_col

Name of the existing column in the edges data.frame that will be used to generate the segment proportional influence (segment PI) for each line feature, in character format.

segpi_col

Name of the new column created in edges that will store the new segment proportional influence values for each feature, in character format.

afv_col

Name of the new column created in edges to store the additive function value for each feature, in character format.

save_local

Logical indicating whether the updated edges should be saved to lsn_path in geopackage format. Defaults to TRUE.

overwrite

A logical indicating whether results should be overwritten if segpi_col and/or afv_col already exists in edges, or edges.gpkg already exists in lsn_path and save_local = TRUE. Default = TRUE.

Value

An sf object representing edges in the LSN, with new segpi_col and afv_col columns. If save_local = TRUE, the updated version of edges will be saved as edges.gpkg in lsn_path.